feat: create act runner Docker image
Create a slimmed-down Docker image to run act tests on GitHub runners.
The nektos/act-environments-ubuntu:18.04 image was used to run tests
locally, but GitHub-hosted runners only have 14 GB of storage space
available. Because the image is over 18 GB, the tests fail when pulling
the image.

This image contains only the packages required to run the
action-taskcat tests, making it small enough to run on the GitHub
runners.

Associated issue: ShahradR/action-taskcat#14